requirements.txt
# MADORO CODE Dependencies

# GUI
PyQt6>=6.4.0

# Core
requests>=2.28.0
PyYAML>=6.0

# API Clients (install the ones you need)
openai>=1.0.0              # For DeepSeek API
anthropic>=0.18.0          # For Claude API
google-generativeai>=0.5.0 # For Gemini API

# Build (optional, for creating executable)
pyinstaller>=5.0
